backend/app/requirements.txt
bson
chardet
pip
setuptools>=68.0.0
wheel>=0.41.0
flask==3.0.0
genai
langchain-google-genai
requests
playwright
python-dotenv
google-generativeai
psycopg2-binary
pinecone
langchain
langchain-community
langchain-core
langchain-pinecone
langchain-text-splitters
mysql-connector-python
beautifulsoup4
lxml
unstructured
html2text
nltk
pandas
openpyxl